pyproject.toml
[project]
name = "flowers"
version = "1.0.0"
authors = [{ name = "Till Muser", email = "musertill@gmail.com" }]
license = {file = "LICENSE"}
readme = {file = "README.md"}
description = "A Warp Drive for Neural PDE Solvers"

requires-python = "~=3.13.0"
dependencies = [
    "torch >= 2.0.1",
    "torchvision >= 0.15.2",
    "numpy < 2.0.0",
    "transformers == 4.55.0",
    "matplotlib",
    "accelerate>=0.32.0",
    "wandb==0.22.2",
    "h5py",
    "pandas",
    "pyyaml",
    "netcdf4>=1.7.2",
    "einops>=0.8.1",
    "scipy>=1.16.1",
    "pytorch-lightning>=2.3.3",
    "ninja>=1.13.0",
    "ipykernel>=6.30.1",
    "seaborn>=0.13.2",
    "huggingface-hub[cli]>=0.34.4",
    "xarray>=2025.8.0",
    "torchinfo>=1.8.0",
    "the-well>=1.1.0",
    "hydra-core>=1.3.2",
    "denoising-diffusion-pytorch>=2.2.5",
    "timm>=1.0.20",
    "neuraloperator>=2.0.0",
    "triton>=3.4.0",
    "ruff>=0.14.5",
    "vape4d>=0.2.7",
    "tqdm>=4.67.1",
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[[tool.uv.index]]
name = "pyg"
url = "https://data.pyg.org/whl/torch-2.7.0+cu126.html"
ignore-error-codes = [403]
format = "flat"

[tool.pyright]
exclude = ["data", "checkpoints", "logs", "outputs"]
typeCheckingMode = "off"
